moved traceToFlamebeaver to flamegraph package and add trace property to FramegraphRenderer
